pyproject.toml
[tool.poetry]
name = "pypelines"
version = "0.1.0"
description = ""
authors = ["James Roseman <james.roseman@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
argparse = "^1.4.0"
python = "^3.10"
toml = "^0.10.2"
pyyaml = "^6.0.1"
jinja2 = "^3.1.4"
click = "^8.1.7"
rich = "^13.7.1"
apache-beam = "^2.57.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"

[tool.poetry.scripts]
pypelines = "pypelines.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
